requirements.txt
requests
beautifulsoup4
openai
gspread
oauth2client
python-dotenv
